made main entry file as jsx
