+++ .eslintignore
@@ -1,1 +1,3 @@
 app/libs/*
+app/scss/*
+app/Components/TileExporter/MapSpells.js
